Reduce panel width by 1px in partial strut
This makes the strut do the right thing in multi-monitor setups, because
pixel counts are inclusive, so setting the end position to the screen width
results in a strut which is that width plus one (pixel at row zero). That
made window maximization do not cover the top area of the screen in the
second monitor. Reducing 1px from the total width does the right thing.