Do not set the hint for skipping the taskbar
By disabling the window manager hint to skip taskbar, it will still not be
shown in a "regular" taskbar because of the dock hint, but the panel will
show up when enumerating dock-type windows. For example OpenBox supports
this by using Ctrl-Alt-Tab to cycle windows.

Also, set window name property as GGTraybar.